Cambios para pruebas de consumo de Api

+++ InventarioApp/app/src/main/java/ar/edu/utn/frba/inventario/api/repository/ShipmentRepository.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 
 class ShipmentRepository@Inject constructor(
     private val apiService: ApiService):Repository() {
-    suspend fun getShipment(idShipment: Int): NetworkResult<ShipmentResponse> {
+    suspend fun getShipment(idShipment: Long): NetworkResult<ShipmentResponse> {
         return safeApiCall {apiService.getShipment(idShipment)}
     }
 }
